project/requirements.txt
astroid==2.0.4
bitstring==3.1.5
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.1.1
pyserial==3.4
six==1.11.0
wrapt==1.10.11
